feat: refactor client + fix test
