+++ src/components/common/container/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/utils';
 import { Box } from '../box';
 
 /**
