+++ src/components/common/text-field/index.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 import Image from 'next/image';
 import { Box } from '@/components';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/utils';
 import { useFormContext, Controller, RegisterOptions } from 'react-hook-form';
 import { Typography } from '../typography';
