+++ src/components/common/typography/index.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/utils';
 import type { TypographyProps } from './typography.types';
 
 /**
